+++ pkg/storage/dtos/IcecreamHasIngredients.go
@@ -1,6 +1,6 @@
 package dtos
 
 type IcecreamHasIngredients struct {
-	IcecreamProductId int `db:"IcecreamProductId"`
-	IngredientsId     int `db:"IngredientsId"`
+	IcecreamProductId int64 `db:"IcecreamProductId"`
+	IngredientsId     int64 `db:"IngredientsId"`
 }
