+++ pkg/storage/repos/IcecreamHasSourcingValuesRepo.go
@@ -16,7 +16,7 @@ func NewIcecreamHasSourcingValuesRepo(db storage.Database) *IcecreamHasSourcingV
 	}
 }
 
-func (r *IcecreamHasSourcingValuesRepo) Create(productId int, sourcingValueIds []int) error {
+func (r *IcecreamHasSourcingValuesRepo) Create(productId int64, sourcingValueIds []int64) error {
 	stmt, err := r.db.DB().Preparex(fmt.Sprintf(`
 		INSERT INTO %s.icecream_has_sourcing_values 
 			(icecream_product_id, sourcing_values_id)
